feat: implement auto-save functionality and related hooks
- Add useAutoSave hook for debounced saving of presentation state to the backend.
- Create useSlideMutations hook to handle slide updates and trigger auto-save on changes.
- Introduce useToast hook for displaying success/error messages.
- Implement useZoomFullscreen hook for managing zoom levels and fullscreen mode.
- Add editorDefaults utility for managing default presentation metadata and slides.
- Enhance session storage handling for presentations with readStoredPresentation and writeStoredPresentation functions.
